feat: use browser context #138
This feature uses puppeteer.BrowserContext for the convertion tasks.
Now, not for each task, an own browser instance is instantiated, but
only one. This one and only is used, and only browserContexts are passed
down to the convertion functions.
Reduces CPU Load, eliminates EventEmitter warnings and speed up
conversion of large amount of files by ~50%.