Do not emit process title to stdout
Redirecting stdout into a target file is a common cli operation.
Emitting the process title to stdout prepends a few bytes before
the standard %PDF-#.# comment.

Most readers seem to ignore these bytes, but it's common for
some tools to read a few bytes to determine if a file is a
PDF.